requirements.txt
PyQt6==6.9.0
send2trash==1.8.0
Pillow==10.4.0
openai==1.40.0